chore: move endpoint to new file
