+++ src/backend/base/langflow/api/router.py
@@ -14,6 +14,7 @@ from langflow.api.v1 import (
     validate_router,
     variables_router,
     folders_router,
+    starter_projects_router,
 )
 
 router = APIRouter(
@@ -31,3 +32,4 @@ router.include_router(variables_router)
 router.include_router(files_router)
 router.include_router(monitor_router)
 router.include_router(folders_router)
+router.include_router(starter_projects_router)
