+++ src/backend/base/langflow/api/v1/__init__.py
@@ -10,6 +10,7 @@ from langflow.api.v1.users import router as users_router
 from langflow.api.v1.validate import router as validate_router
 from langflow.api.v1.variable import router as variables_router
 from langflow.api.v1.folders import router as folders_router
+from langflow.api.v1.starter_projects import router as starter_projects_router
 
 __all__ = [
     "chat_router",
@@ -24,4 +25,5 @@ __all__ = [
     "monitor_router",
     "files_router",
     "folders_router",
+    "starter_projects_router",
 ]
